libhwcomposer: Retain support for legacy graphics (3/4)
This fixes errors when booting. If triple buffering is enabled
on a 2.6.35 kernel with an irregular or old graphics system
(such as an MSM7x27 phone) then it will give errors about
surfaceflinger not finding the file or directory and hence
enters a nonstop bootloop in which it tries again and again to
find and open the gralloc, hwcomposer and copybit libraries.

To use this fix, insert the following lines in your BoardConfig.mk:
COMMON_GLOBAL_CFLAGS += -DUSES_LEGACY_GRAPHICS
BOARD_USES_LEGACY_GRAPHICS := true

Change-Id: I3173fb7f90918d6e891e3263b6e5e37d98cb219c